requirements.txt
aiogram==2.19
python-dotenv==0.20.0
requests>=2.22.0
schedule~=1.1.0